requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
python-dotenv==1.0.1

requests~=2.32.3
faiss-cpu

numpy~=1.24.4

beautifulsoup4~=4.13.4

scikit-learn~=1.3.2
